requirements.txt
faiss-cpu
groq
langchain-groq
langchain_google_genai
langchain
langchain_community
python-dotenv
pypdf
google-cloud-aiplatform>=1.38
flask
PyPDF2
nltk
spacy
flashtext
pke
numpy
bert-extractive-summarizer
flask-sqlalchemy
flask_mail
langchain
chromadb
urllib3
PyMuPDF
python-dotenv
unstructured
extract-msg
tabulate
pandoc
pypandoc
tqdm
sentence_transformers
langchain-community 
tiktoken 
langchainhub 
langchain-cohere
pymupdf
chroma-migrate
langchain 
langchain-experimental 